Update Java doc to reflect recent image changes
There are now four Java images (matrix of RHEL7/RHEL8 and JDK8/JDK11).
Update the blurb and reference to JDK versions accordingly.

Update the version of the included Jolokia and Maven versions.

Add a further three example "docker pull" commands for the other image
variants. Their names are hopefully self explanatory wrt RHEL and Java
version (the original one is the least clear from the name)

Update the default Maven arguments example and the list of environment
variables for configuring Maven in an S2I build. There are actually
many more variables that can be used, but I'm not sure whether this
document should be comprehensive.

ARTIFACT_DIR was deprecated in favour of MAVEN_S2I_ARTIFACT_DIRS.
Update the final examples to reflect this.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>